backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
aptos-sdk
requests
pydantic
